[EDF-0] OAuth not working. Add other screens for now.
