+++ RD.CanMusicMakeYouRunFaster/AndroidApp/AndroidApp.csproj
@@ -63,7 +63,7 @@
   <ItemGroup>
     <Compile Include="Resources\Resource.designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SpotifyAuthActivity.cs" />
+    <Compile Include="StravaAuthActivity.cs" />
     <Compile Include="WelcomeActivity.cs" />
   </ItemGroup>
   <ItemGroup>
