+++ RD.CanMusicMakeYouRunFaster/AndroidApp/Resources/layout/ConnectRunningServices.xml
@@ -16,5 +16,5 @@
         android:textAppearance="?android:attr/textAppearanceLarge"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:id="@+id/textView1" />
+        android:id="@+id/runningInfoText" />
 </RelativeLayout>
