--- RD.CanMusicMakeYouRunFaster/AndroidApp/SpotifyAuthActivity.cs
@@ -1,28 +1,0 @@
-﻿namespace AndroidApp
-{
-    using Android.App;
-    using Android.Content;
-    using Android.Content.PM;
-    using Android.OS;
-    using Android.Runtime;
-    using Android.Views;
-    using Android.Widget;
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-
-    [Activity(NoHistory = true, LaunchMode = LaunchMode.SingleTop)]
-    [IntentFilter(new[] { Android.Content.Intent.ActionView },
-    Categories = new[] { Android.Content.Intent.CategoryDefault, Android.Content.Intent.CategoryBrowsable },
-    DataScheme = "myapp")]
-    public class SpotifyAuthActivity : Xamarin.Essentials.WebAuthenticatorCallbackActivity
-    {
-        protected override void OnCreate(Bundle savedInstanceState)
-        {
-            base.OnCreate(savedInstanceState);
-
-            // Create your application here
-        }
-    }
-}
